+++ PHP/JSON-RPC/config_groups.php
@@ -1,0 +1,15 @@
+<?php
+
+//alt id by groups
+
+$GROUP_ALT_ID['56a9da979a0bf84c09c316ca'] = 0; //SS
+$GROUP_ALT_ID['56a9da979a0bf84c09c316c4'] = 1; //oo au
+$GROUP_ALT_ID['56a9da979a0bf84c09c316d0'] = 4; //optix
+$GROUP_ALT_ID['56a9da979a0bf84c09c316dc'] = 0; //ss nz
+$GROUP_ALT_ID['56a9da979a0bf84c09c316e2'] = 2; //oo ca
+$GROUP_ALT_ID['56a9da979a0bf84c09c316d6'] = 3; //oo de
+$GROUP_ALT_ID['56a9da979a0bf84c09c316ee'] = 1000; //ss dis
+$GROUP_ALT_ID['56a9da979a0bf84c09c316f4'] = 1000; //ss gen
+$GROUP_ALT_ID['56a9da979a0bf84c09c316fa'] = 0; //ss callcenter
+$GROUP_ALT_ID['56a9da979a0bf84c09c31700'] = 5; //drug
+$GROUP_ALT_ID['56a9da979a0bf84c09c31700'] = 5; //drug callcenter